Moving search result to component.
